# chore: Speed up random poly generation