--- a/halo2_proofs/Cargo.toml
+++ b/halo2_proofs/Cargo.toml
@@ -52,6 +52,7 @@ rand_core = { version = "0.6", default-features = false }
 tracing = "0.1"
 blake2b_simd = "1"
 maybe-rayon = {version = "0.1.0", default-features = false}
+rand_chacha = { version = "0.3", optional = true }
 
 # Developer tooling dependencies
 plotters = { version = "0.3.0", default-features = false, optional = true }
@@ -69,7 +70,7 @@ getrandom = { version = "0.2", features = ["js"] }
 
 [features]
 default = ["batch", "multicore"]
-multicore = ["maybe-rayon/threads"]
+multicore = ["maybe-rayon/threads", "rand_chacha"]
 dev-graph = ["plotters", "tabbycat"]
 test-dev-graph = [
     "dev-graph",